+++ MLEM.Extended/Font/GenericBitmapFont.cs
@@ -32,14 +32,14 @@ namespace MLEM.Extended.Font {
         }
 
         /// <inheritdoc />
-        protected override float MeasureChar(char c) {
-            var region = this.Font.GetCharacterRegion(c);
+        protected override float MeasureCharacter(int codePoint) {
+            var region = this.Font.GetCharacterRegion(codePoint);
             return region != null ? new Vector2(region.XAdvance, region.Height).X : 0;
         }
 
         /// <inheritdoc />
-        protected override void DrawChar(SpriteBatch batch, string cString, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects effects, float layerDepth) {
-            batch.DrawString(this.Font, cString, position, color, rotation, Vector2.Zero, scale, effects, layerDepth);
+        protected override void DrawCharacter(SpriteBatch batch, int codePoint, string character, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects effects, float layerDepth) {
+            batch.DrawString(this.Font, character, position, color, rotation, Vector2.Zero, scale, effects, layerDepth);
         }
 
     }