+++ MLEM/Font/GenericFont.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Text;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using MLEM.Extensions;
 using MLEM.Misc;
 
 namespace MLEM.Font {
@@ -50,35 +49,36 @@ namespace MLEM.Font {
         public abstract float LineHeight { get; }
 
         /// <summary>
-        /// Measures the width of the given character with the default scale for use in <see cref="MeasureString(string,bool)"/>.
+        /// Measures the width of the given code point with the default scale for use in <see cref="MeasureString(string,bool)"/>.
         /// Note that this method does not support <see cref="Nbsp"/>, <see cref="Zwsp"/> and <see cref="Emsp"/> for most generic fonts, which is why <see cref="MeasureString(string,bool)"/> should be used even for single characters.
         /// </summary>
-        /// <param name="c">The character whose width to calculate</param>
+        /// <param name="codePoint">The code point whose width to calculate</param>
         /// <returns>The width of the given character with the default scale</returns>
-        protected abstract float MeasureChar(char c);
+        protected abstract float MeasureCharacter(int codePoint);
 
         /// <summary>
-        /// Draws the given character with the given data for use in <see cref="DrawString(Microsoft.Xna.Framework.Graphics.SpriteBatch,System.Text.StringBuilder,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Color,float,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Graphics.SpriteEffects,float)"/>.
+        /// Draws the given code point with the given data for use in <see cref="DrawString(Microsoft.Xna.Framework.Graphics.SpriteBatch,System.Text.StringBuilder,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Color,float,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Vector2,Microsoft.Xna.Framework.Graphics.SpriteEffects,float)"/>.
         /// Note that this method is only called internally.
         /// </summary>
         /// <param name="batch">The sprite batch to draw with.</param>
-        /// <param name="cString">A string representation of the character which will be drawn.</param>
+        /// <param name="codePoint">The code point which will be drawn.</param>
+        /// <param name="character">A string representation of the character which will be drawn.</param>
         /// <param name="position">The drawing location on screen.</param>
         /// <param name="color">A color mask.</param>
         /// <param name="rotation">A rotation of this character.</param>
         /// <param name="scale">A scaling of this character.</param>
         /// <param name="effects">Modificators for drawing. Can be combined.</param>
         /// <param name="layerDepth">A depth of the layer of this character.</param>
-        protected abstract void DrawChar(SpriteBatch batch, string cString, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects effects, float layerDepth);
+        protected abstract void DrawCharacter(SpriteBatch batch, int codePoint, string character, Vector2 position, Color color, float rotation, Vector2 scale, SpriteEffects effects, float layerDepth);
 
         ///<inheritdoc cref="SpriteBatch.DrawString(SpriteFont,string,Vector2,Color,float,Vector2,float,SpriteEffects,float)"/>
         public void DrawString(SpriteBatch batch, string text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
-            this.DrawString(batch, new CharSource(text), position, color, rotation, origin, scale, effects, layerDepth);
+            this.DrawString(batch, new CodePointSource(text), position, color, rotation, origin, scale, effects, layerDepth);
         }
 
         ///<inheritdoc cref="SpriteBatch.DrawString(SpriteFont,string,Vector2,Color,float,Vector2,float,SpriteEffects,float)"/>
         public void DrawString(SpriteBatch batch, StringBuilder text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
-            this.DrawString(batch, new CharSource(text), position, color, rotation, origin, scale, effects, layerDepth);
+            this.DrawString(batch, new CodePointSource(text), position, color, rotation, origin, scale, effects, layerDepth);
         }
 
         ///<inheritdoc cref="SpriteBatch.DrawString(SpriteFont,string,Vector2,Color,float,Vector2,float,SpriteEffects,float)"/>
@@ -103,19 +103,19 @@ namespace MLEM.Font {
 
         /// <summary>
         /// Measures the width of the given string when drawn with this font's underlying font.
-        /// This method uses <see cref="MeasureChar"/> internally to calculate the size of known characters and calculates additional characters like <see cref="Nbsp"/>, <see cref="Zwsp"/> and <see cref="Emsp"/>.
+        /// This method uses <see cref="MeasureCharacter"/> internally to calculate the size of known characters and calculates additional characters like <see cref="Nbsp"/>, <see cref="Zwsp"/> and <see cref="Emsp"/>.
         /// If the text contains newline characters (\n), the size returned will represent a rectangle that encompasses the width of the longest line and the string's full height.
         /// </summary>
         /// <param name="text">The text whose size to calculate</param>
         /// <param name="ignoreTrailingSpaces">Whether trailing whitespace should be ignored in the returned size, causing the end of each line to be effectively trimmed</param>
         /// <returns>The size of the string when drawn with this font</returns>
         public Vector2 MeasureString(string text, bool ignoreTrailingSpaces = false) {
-            return this.MeasureString(new CharSource(text), ignoreTrailingSpaces, null);
+            return this.MeasureString(new CodePointSource(text), ignoreTrailingSpaces, null);
         }
 
         /// <inheritdoc cref="MeasureString(string,bool)"/>
         public Vector2 MeasureString(StringBuilder text, bool ignoreTrailingSpaces = false) {
-            return this.MeasureString(new CharSource(text), ignoreTrailingSpaces, null);
+            return this.MeasureString(new CodePointSource(text), ignoreTrailingSpaces, null);
         }
 
         /// <summary>
@@ -129,12 +129,12 @@ namespace MLEM.Font {
         /// <param name="ellipsis">The characters to add to the end of the string if it is too long</param>
         /// <returns>The truncated string, or the same string if it is shorter than the maximum width</returns>
         public string TruncateString(string text, float width, float scale, bool fromBack = false, string ellipsis = "") {
-            return this.TruncateString(new CharSource(text), width, scale, fromBack, ellipsis, null).ToString();
+            return this.TruncateString(new CodePointSource(text), width, scale, fromBack, ellipsis, null).ToString();
         }
 
         /// <inheritdoc cref="TruncateString(string,float,float,bool,string)"/>
         public StringBuilder TruncateString(StringBuilder text, float width, float scale, bool fromBack = false, string ellipsis = "") {
-            return this.TruncateString(new CharSource(text), width, scale, fromBack, ellipsis, null);
+            return this.TruncateString(new CodePointSource(text), width, scale, fromBack, ellipsis, null);
         }
 
         /// <summary>
@@ -165,22 +165,24 @@ namespace MLEM.Font {
         /// <param name="scale">The scale to use for width measurements</param>
         /// <returns>The split string as an enumerable of split sections</returns>
         public IEnumerable<string> SplitStringSeparate(string text, float width, float scale) {
-            return this.SplitStringSeparate(new CharSource(text), width, scale, null);
+            return this.SplitStringSeparate(new CodePointSource(text), width, scale, null);
         }
 
         /// <inheritdoc cref="SplitStringSeparate(string,float,float)"/>
         public IEnumerable<string> SplitStringSeparate(StringBuilder text, float width, float scale) {
-            return this.SplitStringSeparate(new CharSource(text), width, scale, null);
+            return this.SplitStringSeparate(new CodePointSource(text), width, scale, null);
         }
 
-        internal Vector2 MeasureString(CharSource text, bool ignoreTrailingSpaces, Func<int, GenericFont> fontFunction) {
+        internal Vector2 MeasureString(CodePointSource text, bool ignoreTrailingSpaces, Func<int, GenericFont> fontFunction) {
             var size = Vector2.Zero;
             if (text.Length <= 0)
                 return size;
             var xOffset = 0F;
-            for (var i = 0; i < text.Length; i++) {
-                var font = fontFunction?.Invoke(i) ?? this;
-                switch (text[i]) {
+            var index = 0;
+            while (index < text.Length) {
+                var font = fontFunction?.Invoke(index) ?? this;
+                var (codePoint, length) = text.GetCodePoint(index);
+                switch (codePoint) {
                     case '\n':
                         xOffset = 0;
                         size.Y += this.LineHeight;
@@ -189,74 +191,80 @@ namespace MLEM.Font {
                         xOffset += this.LineHeight;
                         break;
                     case GenericFont.Nbsp:
-                        xOffset += font.MeasureChar(' ');
+                        xOffset += font.MeasureCharacter(' ');
                         break;
                     case GenericFont.Zwsp:
                         // don't add width for a zero-width space
                         break;
                     case ' ':
-                        if (ignoreTrailingSpaces && GenericFont.IsTrailingSpace(text, i)) {
+                        if (ignoreTrailingSpaces && GenericFont.IsTrailingSpace(text, index)) {
                             // if this is a trailing space, we can skip remaining spaces too
-                            i = text.Length - 1;
+                            index = text.Length - 1;
                             break;
                         }
-                        xOffset += font.MeasureChar(' ');
+                        xOffset += font.MeasureCharacter(' ');
                         break;
                     default:
-                        xOffset += font.MeasureChar(text[i]);
+                        xOffset += font.MeasureCharacter(codePoint);
                         break;
                 }
                 // increase x size if this line is the longest
                 if (xOffset > size.X)
                     size.X = xOffset;
+                index += length;
             }
             // include the last line's height too!
             size.Y += this.LineHeight;
             return size;
         }
 
-        internal StringBuilder TruncateString(CharSource text, float width, float scale, bool fromBack, string ellipsis, Func<int, GenericFont> fontFunction) {
+        internal StringBuilder TruncateString(CodePointSource text, float width, float scale, bool fromBack, string ellipsis, Func<int, GenericFont> fontFunction) {
             var total = new StringBuilder();
-            for (var i = 0; i < text.Length; i++) {
+            var index = 0;
+            while (index < text.Length) {
+                var innerIndex = fromBack ? text.Length - 1 - index : index;
+                var (codePoint, length) = text.GetCodePoint(innerIndex, fromBack);
                 if (fromBack) {
-                    total.Insert(0, text[text.Length - 1 - i]);
+                    total.Insert(0, char.ConvertFromUtf32(codePoint));
                 } else {
-                    total.Append(text[i]);
+                    total.Append(char.ConvertFromUtf32(codePoint));
                 }
 
-                var font = fontFunction?.Invoke(i) ?? this;
-                if (font.MeasureString(total + ellipsis).X * scale >= width) {
+                if (this.MeasureString(new CodePointSource(total + ellipsis), false, fontFunction).X * scale >= width) {
                     if (fromBack) {
-                        return total.Remove(0, 1).Insert(0, ellipsis);
+                        return total.Remove(0, length).Insert(0, ellipsis);
                     } else {
-                        return total.Remove(total.Length - 1, 1).Append(ellipsis);
+                        return total.Remove(total.Length - length, length).Append(ellipsis);
                     }
                 }
+                index += length;
             }
             return total;
         }
 
-        internal IEnumerable<string> SplitStringSeparate(CharSource text, float width, float scale, Func<int, GenericFont> fontFunction) {
+        internal IEnumerable<string> SplitStringSeparate(CodePointSource text, float width, float scale, Func<int, GenericFont> fontFunction) {
             var currWidth = 0F;
             var lastSpaceIndex = -1;
             var widthSinceLastSpace = 0F;
             var curr = new StringBuilder();
-            for (var i = 0; i < text.Length; i++) {
-                var c = text[i];
-                if (c == '\n') {
+            var index = 0;
+            while (index < text.Length) {
+                var (codePoint, length) = text.GetCodePoint(index);
+                if (codePoint == '\n') {
                     // fake split at pre-defined new lines
-                    curr.Append(c);
+                    curr.Append('\n');
                     lastSpaceIndex = -1;
                     widthSinceLastSpace = 0;
                     currWidth = 0;
                 } else {
-                    var font = fontFunction?.Invoke(i) ?? this;
-                    var cWidth = font.MeasureString(c.ToCachedString()).X * scale;
-                    if (c == ' ' || c == GenericFont.Emsp || c == GenericFont.Zwsp) {
+                    var font = fontFunction?.Invoke(index) ?? this;
+                    var character = char.ConvertFromUtf32(codePoint);
+                    var charWidth = font.MeasureString(character).X * scale;
+                    if (codePoint == ' ' || codePoint == GenericFont.Emsp || codePoint == GenericFont.Zwsp) {
                         // remember the location of this (breaking!) space
                         lastSpaceIndex = curr.Length;
                         widthSinceLastSpace = 0;
-                    } else if (currWidth + cWidth >= width) {
+                    } else if (currWidth + charWidth >= width) {
                         // check if this line contains a space
                         if (lastSpaceIndex < 0) {
                             // if there is no last space, the word is longer than a line so we split here
@@ -275,16 +283,17 @@ namespace MLEM.Font {
                     }
 
                     // add current character
-                    currWidth += cWidth;
-                    widthSinceLastSpace += cWidth;
-                    curr.Append(c);
+                    currWidth += charWidth;
+                    widthSinceLastSpace += charWidth;
+                    curr.Append(character);
                 }
+                index += length;
             }
             if (curr.Length > 0)
                 yield return curr.ToString();
         }
 
-        private void DrawString(SpriteBatch batch, CharSource text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
+        private void DrawString(SpriteBatch batch, CodePointSource text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
             var (flipX, flipY) = (0F, 0F);
             var flippedV = (effects & SpriteEffects.FlipVertically) != 0;
             var flippedH = (effects & SpriteEffects.FlipHorizontally) != 0;
@@ -318,55 +327,38 @@ namespace MLEM.Font {
             }
 
             var offset = Vector2.Zero;
-            for (var i = 0; i < text.Length; i++) {
-                var c = text[i];
-                if (c == '\n') {
+            var index = 0;
+            while (index < text.Length) {
+                var (codePoint, length) = text.GetCodePoint(index);
+                if (codePoint == '\n') {
                     offset.X = 0;
                     offset.Y += this.LineHeight;
-                    continue;
+                } else {
+                    var character = char.ConvertFromUtf32(codePoint);
+                    var charSize = this.MeasureString(character);
+
+                    var charPos = offset;
+                    if (flippedH)
+                        charPos.X += charSize.X;
+                    if (flippedV)
+                        charPos.Y += charSize.Y - this.LineHeight;
+                    Vector2.Transform(ref charPos, ref trans, out charPos);
+
+                    this.DrawCharacter(batch, codePoint, character, charPos, color, rotation, scale, effects, layerDepth);
+                    offset.X += charSize.X;
                 }
-
-                var cString = c.ToCachedString();
-                var cSize = this.MeasureString(cString);
-
-                var charPos = offset;
-                if (flippedH)
-                    charPos.X += cSize.X;
-                if (flippedV)
-                    charPos.Y += cSize.Y - this.LineHeight;
-                Vector2.Transform(ref charPos, ref trans, out charPos);
-
-                this.DrawChar(batch, cString, charPos, color, rotation, scale, effects, layerDepth);
-                offset.X += cSize.X;
+                index += length;
             }
         }
 
-        private static bool IsTrailingSpace(CharSource s, int index) {
-            for (var i = index + 1; i < s.Length; i++) {
-                if (s[i] != ' ')
+        private static bool IsTrailingSpace(CodePointSource s, int index) {
+            while (index < s.Length) {
+                var (codePoint, length) = s.GetCodePoint(index);
+                if (codePoint != ' ')
                     return false;
+                index += length;
             }
             return true;
-        }
-
-        internal readonly struct CharSource {
-
-            private readonly string strg;
-            private readonly StringBuilder builder;
-
-            public int Length => this.strg?.Length ?? this.builder.Length;
-            public char this[int index] => this.strg?[index] ?? this.builder[index];
-
-            public CharSource(string strg) {
-                this.strg = strg;
-                this.builder = null;
-            }
-
-            public CharSource(StringBuilder builder) {
-                this.strg = null;
-                this.builder = builder;
-            }
-
         }
 
     }